BOOT:Modify the Linux kernel startup mode
Because nandflash does not support the fat file system, the fat file image is stored in ubifs.
After the fat file system image is loaded to the ramblock device, the content in the ramblock device
is read to start the linux kernel.And we should check mmc device num to decide flash which device
(eMMC or nandflash).

Signed-off-by: laoyening <1337849843@qq.com>